==> 05_plot_graphs.ipynb <==
import xarray as xr
import pandas as pd
import numpy as np
import os 
import pickle
import timeit
import math
import time
import sparse
import dask
import dask.array as da
from dask.diagnostics import ProgressBar
from scipy import interpolate
from pylab import *
from numpy import *
from glob import glob
from os import path
from tqdm import tqdm

import matplotlib.pyplot as plt
import matplotlib as mpl
from matplotlib import ticker, cm, rc
from mpl_toolkits.axes_grid1 import make_axes_locatable
from mpl_toolkits.axes_grid1.inset_locator import inset_axes
from matplotlib.patches import Rectangle
import matplotlib.colors as colors
import matplotlib.cbook as cbook
from matplotlib import cm
import seaborn as sns

import cartopy
import cartopy.crs as ccrs
import cartopy.feature as cfeature
from cartopy.mpl.gridliner import LONGITUDE_FORMATTER, LATITUDE_FORMATTER
cartopy.config['data_dir']= os.getenv('CARTOPY_DIR', cartopy.config.get('data_dir'))
import cartopy.io.shapereader as shpreader
from shapely.geometry import Point

import matplotlib
import matplotlib.gridspec as gridspec
from matplotlib.colors import ListedColormap
from matplotlib.lines import Line2D
from mpl_toolkits.axes_grid1.inset_locator import inset_axes
from matplotlib.ticker import ScalarFormatter
from copy import copy

plt.ion()  # To trigger the interactive inline mode

import warnings
warnings.filterwarnings("ignore", category=RuntimeWarning) 

from ll_Balltree import *
%run -i 'll_Balltree.py'

acc = 4

outputDir = 'data/posterior_computation_data/'
outputgraphsDir = outputDir + 'graphs/'
gom_masks = xr.open_dataset(outputDir + 'gom_masks.nc')

# GLOBAL CONSTANTS
MIN_LON = np.min(gom_masks['lon'].values)
MAX_LON = np.max(gom_masks['lon'].values)
MIN_LAT = np.min(gom_masks['lat'].values)
MAX_LAT = np.max(gom_masks['lat'].values)

#domain width and height (cell counts)
LAT_SIZE = gom_masks.dims['lat']
LON_SIZE = gom_masks.dims['lon']

#cell size
D_LON = gom_masks["lon"][1].values - gom_masks["lon"][0].values
D_LAT = gom_masks["lat"][1].values - gom_masks["lat"][0].values

BIN_CELL_LATS = gom_masks.bin_cell_lats.values
BIN_CELL_LONS = gom_masks.bin_cell_lons.values

print("Rectangular Boundaries (lons, lats):", (np.round(MIN_LON, acc), np.round(MAX_LON, acc)), ',',
                                             (np.round(MIN_LAT, acc), np.round(MAX_LAT, acc)))
print("cell_size (width (dlon),height (dlat)): ", np.round(D_LON, acc), np.round(D_LAT, acc))

# Load domain_cell_tree
fileObj = open(outputDir + 'output_dict.obj', 'rb')
output_dict = pickle.load(fileObj)
fileObj.close()

## n_cell_beaching = output['n_cell_beaching']
n_cell_source = output_dict['n_cell_source']
n_window_beaching = output_dict['n_window_beaching'] 
n_window_source = output_dict['n_window_source']
n_cell_beaching = output_dict['n_cell_beaching']

source_cell_mask = output_dict['source_cell_mask']

print('Number of Beaching Cells:', n_cell_beaching)
print()
print('Number of Source Cells:', n_cell_source)
print('        Number of Coastal Cells:', int(np.sum(gom_masks.coastalmask.values[source_cell_mask])))
print('        Number of Ocean Cells:', int(np.sum(gom_masks.oceanmask.values[source_cell_mask])))
print()
print('Number of Beaching Windows:', n_window_beaching)
print('Number of Source Windows:', n_window_source)

category_colors = sns.color_palette("tab20", 15, desat = 0.7)

# Load agg values
fileObj = open(outputDir + 'agg_dict.obj', 'rb')
output_agg = pickle.load(fileObj)
fileObj.close()

group_source_region_names = output_agg['group_source_region_names']

# Assign Colors
region_colors = dict(zip(group_source_region_names, category_colors))

# Load All Data

#############
Figure_11_l_raw = xr.load_dataset(outputDir + 'Figure_11_l_raw.nc')
Figure_11_l_vals = Figure_11_l_raw.l_n_post.values 

Figure_11_l = np.zeros((LAT_SIZE, LON_SIZE))
Figure_11_l[source_cell_mask] = Figure_11_l_vals *100
#############

#############
Figure_11_r_raw = xr.load_dataset(outputDir + 'Figure_11_r_raw.nc')
Figure_11_r_vals = Figure_11_r_raw.r_n_post.values

Figure_11_r = np.zeros((LAT_SIZE, LON_SIZE))
Figure_11_r[source_cell_mask] = Figure_11_r_vals *100
#############

#############
Figure_11_f_raw = xr.load_dataset(outputDir + 'Figure_11_f_raw.nc')
Figure_11_f_vals = Figure_11_f_raw.f_n_post.values

Figure_11_f = np.zeros((LAT_SIZE, LON_SIZE))
Figure_11_f[source_cell_mask] = Figure_11_f_vals*100
#############

#############
Figure_11_raw = xr.load_dataset(outputDir + 'Fig11_greyscale.nc')
Figure_11_vals = Figure_11_raw.__xarray_dataarray_variable__.values

Figure_11 = np.zeros((LAT_SIZE, LON_SIZE))
Figure_11[source_cell_mask] = Figure_11_vals*100
#############

Fig9 = xr.load_dataset(outputDir + 'Fig9.nc')
Fig9_vals = Fig9.__xarray_dataarray_variable__.values*100
Fig9_gs_region = (Fig9.group_source_region.values.astype(int) - 1)
Fig9_gb_region = (Fig9.group_beaching_region.values.astype(int) - 1)

Fig13 = xr.load_dataset(outputDir + 'Fig13.nc')
Fig13_vals = Fig13.__xarray_dataarray_variable__.values*100

prior = xr.open_dataset(outputDir + 'prior.nc')

l_prior = prior.l_prior.values * 100
r_prior = prior.r_prior.values * 100
f_prior = prior.f_prior.values * 100

# Galveston Bay (29.34, -94.56)
# New Orleans (29.58, -90.04)
# Tampa (27.56 -82.27)
#Corpus Christi (27.44, 97.24)
# Mérida, Yucatán (20.58, -89.37) 
# Frontera, Tabasco (18.32 -92.38)

# Grijalva River
# Pearl River

major_city_lats = [29.4, 27.44, 29.98, 27.56, 20.58]
major_city_lons = [-94.68, -97.21, -90.15, -82.27, -89.37]
major_city_names = ['Galveston Bay', 'Corpus Christi', 'New Orleans', 'Tampa', 'Mérida']
major_city_hl = ['right', 'right', 'right', 'left','left']
major_city_vl = ['bottom', 'bottom', 'bottom', 'bottom','top']

major_city_df = { 'major_city_names': major_city_names, 'lats': major_city_lats, 'lons': major_city_lons, 
                 'major_city_hl': major_city_hl, 'major_city_vl': major_city_vl}
major_city_df = pd.DataFrame(major_city_df)

def gom_plot_annotations(ax = None, depth = False, depth_border_pad = -5, major_cities = True, major_rivers = True, ocean_labels = False):
    
    if major_cities:
        # ANNOTATE Major Cities
        plt.scatter(major_city_df.lons, major_city_df.lats, color = 'black', s = 7, zorder=11)
        # add labels to all points
        for i in range(len(major_city_df)):
            if (major_city_df.major_city_names.iloc[i] == 'New Orleans'):
                plt.text(major_city_df.lons.iloc[i]-0.1, major_city_df.lats.iloc[i]+0.05, 
                     major_city_df.major_city_names.iloc[i], size = 7, rotation_mode =  'anchor',
                     va=major_city_df.major_city_vl.iloc[i], 
                     ha=major_city_df.major_city_hl.iloc[i], zorder=11)
            elif (major_city_df.major_city_names.iloc[i] == 'Galveston Bay'):
                plt.text(major_city_df.lons.iloc[i]-0.11, major_city_df.lats.iloc[i]+0.15, 
                     major_city_df.major_city_names.iloc[i], size = 7, rotation_mode =  'anchor',
                     va=major_city_df.major_city_vl.iloc[i], 
                     ha=major_city_df.major_city_hl.iloc[i], zorder=11)
            elif (major_city_df.major_city_names.iloc[i] == 'Corpus Christi'):
                plt.text(major_city_df.lons.iloc[i]+1.5, major_city_df.lats.iloc[i]+0.1, 
                     major_city_df.major_city_names.iloc[i], size = 7, rotation_mode =  'anchor',
                     va=major_city_df.major_city_vl.iloc[i], 
                     ha=major_city_df.major_city_hl.iloc[i], zorder=11)
            else:
                plt.text(major_city_df.lons.iloc[i]+0.1, major_city_df.lats.iloc[i]-0.05, 
                     major_city_df.major_city_names.iloc[i], size = 7, rotation_mode =  'anchor',
                     va=major_city_df.major_city_vl.iloc[i], 
                     ha=major_city_df.major_city_hl.iloc[i], zorder=11)


    if major_rivers:
        # ANNOTATE Rivers
        # Mississippi River
        ax.annotate("Mississippi \n River",
                    xy=(-89.3, 29.3), xycoords='data',
                    xytext=(-89.0, 28.3), textcoords='data',
                    size=7, va="center", ha="center",
                    arrowprops=dict(arrowstyle="wedge",
                                    connectionstyle="arc3,rad=0.2", color = 'k'), zorder=11)

        # Grijalva River
        ax.annotate("Grijalva \n River",
                    xy=(-92.73, 18.6), xycoords='data',
                    xytext=(-91.9, 19.5), textcoords='data',
                    size=7, va="center", ha="center",
                    arrowprops=dict(arrowstyle="wedge",
                                    connectionstyle="arc3,rad=0.2", color = 'k'), zorder=11)

        # Pearl River
        ax.annotate("Pearl \n River",
                    xy=(-89.6, 30.2), xycoords='data',
                    xytext=(-89.3, 31.3), textcoords='data',
                    size=7, va="center", ha="center",
                    arrowprops=dict(arrowstyle="wedge",
                                    connectionstyle="arc3,rad=-0.2", color = 'k'), zorder=11)
    
    if ocean_labels:
        # ANNOTATE REGIONS
        # LaTeX Shelf
        plt.text(-94.6, 28.1, "LaTeX Shelf", size = 10, style='italic', rotation_mode =  'anchor', zorder=11)

        # nWFS
        plt.text(-88.0, 29.0, "North West \n Florida Shelf", size = 10, style='italic', rotation_mode =  'anchor', zorder=11)

        # sWFS
        plt.text(-84.5, 27.5, "South West \n Florida Shelf", size = 10, style='italic', 
                 rotation_mode =  'anchor',  rotation=-60, zorder=11)

        # Campeche Bay
        plt.text(-95.0, 20.0, "Bay of \n Campeche", size = 10, style='italic', rotation_mode =  'anchor', zorder=11)

        # Atlantic Ocean
        plt.text(-77.9, 28.0, "Atlantic Ocean", size = 10, style='italic', 
                 rotation_mode =  'anchor',  rotation=90, zorder=11)

        # Caribbean Sea
        plt.text(-85.5, 19.7, "Caribbean Sea", size = 10, style='italic', 
                 rotation_mode =  'anchor',  zorder=11)
    
    if depth:
        # DEPTH
        lon_vals, lat_vals = np.meshgrid(gom_masks['lon'], gom_masks['lat'])

        d = ax.contourf(lon_vals, lat_vals, gom_masks.depth,cmap = plt.cm.Blues, alpha = 0.9, zorder=1)

        # Plot colorbars
        axins_d = inset_axes(ax,
                            width="100%",  
                            height="3%",
                            loc='lower center',
                            borderpad=depth_border_pad
                           )

        axins_d.set_title('Depth [meters]', fontsize = 9)
        fig.colorbar(d, cax=axins_d, orientation="horizontal")
        axins_d.minorticks_on()

def plot_gom(z_order = 2, extend_plot = False):
    # plotting settings
    fig = plt.figure(figsize=(9,10), constrained_layout = True)
    ax = plt.axes(projection=ccrs.PlateCarree())
    gl = ax.gridlines(crs=ccrs.PlateCarree(), draw_labels=True,
                      linewidth=1, color='black', alpha=0.1, linestyle='--')
    gl.top_labels = False
    gl.right_labels = False

    # Geospatial Features
    ax.add_feature(cfeature.LAND, zorder=1,)
    ax.add_feature(cfeature.COASTLINE, linewidth=0.7, zorder=z_order, alpha = 0.3)
    ax.add_feature(cfeature.STATES, linewidth=0.5, zorder=z_order, alpha = 0.5)
    ax.add_feature(cfeature.RIVERS, linewidth=0.7, zorder=z_order,  alpha = 1.0)
    
    if extend_plot:
        ax.set_extent([MIN_LON - 1, MAX_LON, MIN_LAT - 1, MAX_LAT], ccrs.PlateCarree())
    else: 
        ax.set_extent([MIN_LON, MAX_LON, MIN_LAT, MAX_LAT], ccrs.PlateCarree())
    return fig, ax

Overall Proportions

l_prior_prop, r_prior_prop, f_prior_prop = np.sum(l_prior), np.sum(r_prior), np.sum(f_prior)
print('Posterior Proportions | Land', np.round(l_prior_prop, acc), 
                           ' | River',  np.round(r_prior_prop, acc), 
                            '| Fishing ',  np.round(f_prior_prop, acc))

l_post_prop, r_post_prop, f_post_prop = np.sum(Figure_11_l), np.sum(Figure_11_r), np.sum(Figure_11_f)
print('Posterior Proportions | Land', np.round(l_post_prop, acc), 
                           ' | River', np.round(r_post_prop, acc), 
                            '| Fishing ', np.round(f_post_prop, acc))

Blank Map

fig, ax = plot_gom(z_order = 2, extend_plot=False)
    
gom_plot_annotations(ax = ax, depth = True, major_cities = True, major_rivers = True, ocean_labels = True)
#plt.savefig(outputgraphsDir + 'labeled_gom_map.png')
plt.close()

Prior

prior_mpw = xr.open_dataset(outputDir + 'prior_mpw.nc')

land_input = prior_mpw.land_input.mean(dim = 'time').values
river_input = prior_mpw.river_input.mean(dim = 'time').values
fishing_input = prior_mpw.fishing_input.mean(dim = 'time').values

fig, ax = plot_gom(z_order=1, extend_plot=True)

gom_plot_annotations(ax = ax, depth = False, major_cities = True, major_rivers = True, ocean_labels = False)
    
# Custom Features
num_colors = 8
cmap_r = plt.get_cmap('Greys', num_colors)

lon_vals, lat_vals = np.meshgrid(gom_masks['lon'], gom_masks['lat'])

levels_mpw = np.logspace(np.log10(10), np.log10(10_000), 9)
levels_fish = np.logspace(np.log10(1), np.log10(1_000), 9)

# fishing
f = ax.contourf(lon_vals, lat_vals, fishing_input, levels = levels_fish, 
                norm=colors.LogNorm(), extend='both', cmap = plt.cm.Blues, alpha = 1.0, )

# land
within_25km_idx = np.where(~np.isnan(gom_masks.plotcoastidx))
ci25 = np.array(gom_masks.plotcoastidx.values[within_25km_idx], dtype = int)
idx_coast_mask = np.where(gom_masks.coastalmask == 1)
limgm = land_input[idx_coast_mask]
vals_within_25km = np.zeros( (gom_masks.dims['lat'], gom_masks.dims['lon']) ) * np.nan
vals_within_25km[within_25km_idx] = limgm[ci25]

land = ax.contourf(lon_vals, lat_vals, vals_within_25km, levels = levels_mpw, norm=colors.LogNorm(), 
                   extend='both', cmap = plt.cm.Reds,  alpha = 0.8,)

# river
r_lat_idx, r_lon_idx = np.where(river_input > 0.0)
r_Y, r_X, r_Z = gom_masks.lat[r_lat_idx].values, gom_masks.lon[r_lon_idx].values, river_input[r_lat_idx, r_lon_idx] 

df = {'lat': r_Y, 'lon': r_X, 'val': r_Z}
df = pd.DataFrame(df)
df = df.sort_values(by = 'val', ascending = False)

reduced_df = df[df.val >= 1]
reduced_df = reduced_df.sort_values(by = 'val', ascending = True)

river = ax.scatter(reduced_df.lon, reduced_df.lat, c = reduced_df.val, s = 50, 
                   norm=colors.LogNorm(vmin=10, vmax=10_000), alpha = 1.0, cmap = cmap_r, edgecolors='black', linewidths = 0.3)

# Plot colorbars
axins_l = inset_axes(ax,
                    width="100%",  
                    height="3%",
                    loc='lower center',
                    borderpad=-5)

axins_l.set_title('Land MPW [kg/month/km^2]', fontsize = 9)
fig.colorbar(land, cax=axins_l, orientation="horizontal", extend='both', ticks = [10, 100, 1000, 10000])
axins_l.minorticks_on()

axins_r = inset_axes(ax,
                    width="100%",  
                    height="3%",
                    loc='lower center',
                    borderpad=-10)

axins_r.set_title('River MPW  [kg/month]', fontsize = 9)
fig.colorbar(river, cax=axins_r, orientation="horizontal", extend='both', ticks = [1, 10, 100, 1000, 10000])
axins_r.minorticks_on()

axins_f = inset_axes(ax,
                    width="100%",  
                    height="3%",
                    loc='lower center',
                    borderpad=-15)

axins_f.set_title('Fishing Activities [hours/km^2]', fontsize = 9)
fig.colorbar(f, cax=axins_f, orientation="horizontal",extend='both', ticks = [1, 10, 100, 1000])
axins_f.minorticks_on()

#plt.savefig(outputgraphsDir + 'prior_input_map.png')
plt.close()

Figure 9

category_names_fig9 = group_source_region_names[Fig9_gs_region]
# since vals are 0.9999, I have just adjusted them to exactly one, and the plot looks better this way
results_fig9 = {(i+1): Fig9_vals.T[i]/np.sum( Fig9_vals.T[i]) for i in range(np.shape(Fig9_vals)[1])}


def survey(results, category_names):
    """
    Parameters
    ----------
    results : dict
        A mapping from question labels to a list of answers per category.
        It is assumed all lists contain the same number of entries and that
        it matches the length of *category_names*.
    category_names : list of str
        The category labels.
    """
    labels = list(results.keys())
    data = np.array(list(results.values()))
    data_cum = data.cumsum(axis=1)

    fig, ax = plt.subplots(figsize=(9.2, 5))
    ax.invert_yaxis()
    ax.yaxis.set_visible(True)
    ax.set_ylim(0, np.sum(data, axis=1).max())
    ax.set_ylabel("Source Probability %")
    
    ax.set_xlabel('Beaching Region')


    for i, (colname, color) in enumerate(zip(category_names, category_colors)):
        widths = data[:, i]
        starts = data_cum[:, i] - widths
        rects = ax.bar(labels, widths, bottom=starts,width = 0.9, label=colname, color=color, edgecolor='white',)
    legend = ax.legend(ncols=1, bbox_to_anchor=(1.25, 0.5), loc='right', fontsize='medium')

    return fig, ax, legend


fig, ax, legend = survey(results_fig9, category_names_fig9)

labels = group_source_region_names[Fig9_gb_region]
x = [i for i in range(1, len(labels)+1)]

plt.xticks(x, labels, rotation=45)

plt.savefig(outputgraphsDir + 'beaching_region_source_prob.png')
plt.close()

# Texas is the 1 first entry indexed in the dictionary and Mexico the second element in the list of probabilities
print("% of particles sourced in Mexico beaching in Texas: ", np.round(results_fig9[1][1], acc))

Figure 11

fig, ax = plot_gom(z_order = 1, extend_plot=True)

gom_plot_annotations(ax = ax, depth = False, major_cities = True, major_rivers = True, ocean_labels = False)

# Custom Features
num_colors = 8
cmap_r = plt.get_cmap('Greys', num_colors)
lon_vals, lat_vals = np.meshgrid(gom_masks['lon'], gom_masks['lat'])

levels_mpw = np.logspace(np.log10(0.01), np.log10(10), 9)
levels_fish = np.logspace(np.log10(0.001), np.log10(1), 9)


# fishing

f = ax.contourf(lon_vals, lat_vals, Figure_11_f, levels = levels_fish, 
                norm=colors.LogNorm(), extend='both', cmap = plt.cm.Blues, alpha = 1.0, )

# land
within_25km_idx = np.where(~np.isnan(gom_masks.plotcoastidx))
ci25 = np.array(gom_masks.plotcoastidx.values[within_25km_idx], dtype = int)
idx_coast_mask = np.where(gom_masks.coastalmask == 1)
limgm = Figure_11_l[idx_coast_mask]
vals_within_25km = np.zeros( (gom_masks.dims['lat'], gom_masks.dims['lon']) ) * np.nan
vals_within_25km[within_25km_idx] = limgm[ci25]

land = ax.contourf(lon_vals, lat_vals, vals_within_25km, levels = levels_mpw, 
                   norm=colors.LogNorm(), extend='both', cmap = plt.cm.Reds,  alpha = 0.8,)


# river
r_lat_idx, r_lon_idx = np.where(Figure_11_r > 0.0)
r_Y, r_X, r_Z = gom_masks.lat[r_lat_idx].values, gom_masks.lon[r_lon_idx].values, Figure_11_r[r_lat_idx, r_lon_idx] 

# replot top 25 rivers for visibility
df = {'lat': r_Y, 'lon': r_X, 'val': r_Z}
df = pd.DataFrame(df)
df = df.sort_values(by = 'val', ascending = False)
reduced_df = df[df.val >= 0.001]
reduced_df = reduced_df.sort_values(by = 'val', ascending = True)

river = ax.scatter(reduced_df.lon, reduced_df.lat, c = reduced_df.val, s = 50, 
                   norm=colors.LogNorm(vmin=1e-2, vmax=10), alpha = 1.0, cmap = cmap_r, edgecolors='black', linewidths = 0.3)

# Plot colorbars
axins_l = inset_axes(ax,
                    width="100%",  
                    height="3%",
                    loc='lower center',
                    borderpad=-5
                   )

axins_l.set_title('Land probabilities [%]', fontsize = 9)
fig.colorbar(land, cax=axins_l, orientation="horizontal", extend='both', ticks=[0.01, 0.1, 1, 10])
axins_l.minorticks_on()

axins_r = inset_axes(ax,
                    width="100%",  
                    height="3%",
                    loc='lower center',
                    borderpad=-10
                   )

axins_r.set_title('River probabilities  [%]', fontsize = 9)
fig.colorbar(river, cax=axins_r, orientation="horizontal", ticks=[0.01,0.1,1,10], extend='both')
axins_r.minorticks_on()

axins_f = inset_axes(ax,
                    width="100%",  
                    height="3%",
                    loc='lower center',
                    borderpad=-15
                   )

axins_f.set_title('Fishing probabilities [%]', fontsize = 9)
fig.colorbar(f, cax=axins_f, orientation="horizontal",extend='both', ticks=[0.001, 0.01, 0.1, 1])
axins_f.minorticks_on()

plt.savefig(outputgraphsDir + 'posterior_map.png')

plt.close()

top_river = int(np.argmax(Figure_11_r_vals))
tr_x, tr_y =source_cell_mask[0][top_river], source_cell_mask[1][top_river]
print('% Sourced Grijalva River:', np.round(Figure_11_r[tr_x, tr_y], acc))

Figure 13

category_names_fig13 = group_source_region_names
# since vals are 0.999, I have just adjusted them to exactly one, and the plot looks better this way 
results_fig13 = {(i+1): Fig13_vals.T[i]/np.sum(Fig13_vals.T[i]) for i in range(np.shape(Fig13_vals)[1])}


def survey(results, category_names):
    """
    Parameters
    ----------
    results : dict
        A mapping from question labels to a list of answers per category.
        It is assumed all lists contain the same number of entries and that
        it matches the length of *category_names*.
    category_names : list of str
        The category labels.
    """
    labels = list(results.keys())
    data = np.array(list(results.values()))
    data_cum = data.cumsum(axis=1)

    fig, ax = plt.subplots(figsize=(9.2, 5))
    ax.invert_yaxis()
    ax.yaxis.set_visible(True)
    ax.set_ylim(0, np.sum(data, axis=1).max())
    ax.set_ylabel("Source Probability %")
    ax.set_xlabel('Beaching Month')


    for i, (colname, color) in enumerate(zip(category_names, category_colors)):
        widths = data[:, i]
        starts = data_cum[:, i] - widths
        rects = ax.bar(labels, widths, bottom=starts,width = 0.9, label=colname, color=color, edgecolor='white',)

    legend = ax.legend(ncols=1, bbox_to_anchor=(1.25, 0.5), loc='right', fontsize='medium')

    return fig, ax, legend


fig, ax, legend = survey(results_fig13, category_names_fig13)

labels = ['Jan','Feb','Mar','Apr','May','Jun','Jul','Aug','Sept','Oct', 'Nov', 'Dec']
x = [i for i in range(1, len(labels) + 1)]

plt.xticks(x, labels, rotation=45)

plt.savefig(outputgraphsDir + 'beaching_month_source_prob.png')
plt.close()

Nurdle Beaching Location

fig, ax = plot_gom(z_order = 1, extend_plot=False)

gom_plot_annotations(ax = ax, depth = False, major_cities = True, major_rivers = False, ocean_labels = False)
    
# Custom Features
num_colors = 5
cmap_n = plt.get_cmap('BuPu', num_colors)

lon_vals, lat_vals = np.meshgrid(gom_masks['lon'], gom_masks['lat'])


# nurdle
nurdles = gom_masks.nurdlecount.values
r_lat_idx, r_lon_idx = np.where(nurdles > 0.0)
r_Y, r_X, r_Z = gom_masks.lat[r_lat_idx].values, gom_masks.lon[r_lon_idx].values, nurdles[r_lat_idx, r_lon_idx] 

# replot top 25 rivers for visibility
df = {'lat': r_Y, 'lon': r_X, 'val': r_Z}
df = pd.DataFrame(df)
df = df.sort_values(by = 'val', ascending = False)
df = df.sort_values(by = 'val', ascending = True)


river = ax.scatter(df.lon, df.lat, c = df.val, s = 75, 
                   norm=colors.LogNorm(vmin=1, vmax=100000), alpha = 0.7, cmap = cmap_n, edgecolors='black', linewidths = 0.3)


axins_r = inset_axes(ax,
                    width="100%",  
                    height="3%",
                    loc='lower center',
                    borderpad=-5
                   )

axins_r.set_title('Number of Nurdles Collected', fontsize = 9)
fig.colorbar(river, cax=axins_r, orientation="horizontal", ticks=[1, 10, 100,1000, 10000,100000])
axins_r.minorticks_on()

plt.savefig(outputgraphsDir + 'nurdle_map.png')
plt.close()

Nurdle Statistics Table

inputDir_Sim = ('data/output_v2/rechunked/')
inputFiles_Sim = sorted(glob(inputDir_Sim + '*.zarr' ))


outputDir = 'data/posterior_computation_data/'
data_dir = 'data/input/'

def get_nurdles_from_runs(release_year = None, start_month = None, end_month = None):
    ############################################
    # Code directly taken from gom_b_sim_v2.py #
    ############################################
    df_all_nurdles = pd.read_csv(data_dir + 'nurdle_release.csv')
    df_all_nurdles['patrol_date'] = pd.to_datetime(df_all_nurdles['patrol_date'])
    df_all_nurdles.iloc[df_all_nurdles[df_all_nurdles.nurdle_count > 10000].index, 4] = 10000 

    #find the year of release
    df_all_nurdles = df_all_nurdles[(df_all_nurdles.patrol_date.dt.year == release_year) 
                                    & (df_all_nurdles.patrol_date.dt.month >= start_month) 
                                    & (df_all_nurdles.patrol_date.dt.month <= end_month)]
    ############################################
    return df_all_nurdles

# create nurdle dataset
dfs = {}

for filename in inputFiles_Sim: 
    split = filename.split('_')
    month_split = split[4].split('-')
    dfs[filename] = get_nurdles_from_runs(float(split[3]), float(month_split[1]), float(month_split[0]) )
    
df_all_nurdles = pd.concat(dfs).reset_index()


# Load source cell mask
fileObj = open(outputDir + 'output_dict.obj', 'rb')
output_dict = pickle.load(fileObj)
fileObj.close()

source_cell_mask = output_dict['source_cell_mask']

# Load agg values
fileObj = open(outputDir + 'agg_dict.obj', 'rb')
output_agg = pickle.load(fileObj)
fileObj.close()

group_source_region_names = output_agg['group_source_region_names']

# Load gom masks
gom_masks = xr.open_dataset(outputDir + 'gom_masks.nc')

# Create coastal cell tree
idx_coast_mask = np.where(gom_masks.coastalmask == 1)
coast_cells = get_cells(idx_mask = idx_coast_mask, 
                            lats = gom_masks['lat'].values,
                            lons = gom_masks['lon'].values)
coast_cell_tree = build_ll_BallTree(coast_cells)

nurdle_query = get_cells(None, df_all_nurdles.lat.values,  df_all_nurdles.lon.values)
_, cell_beaching_inds = query_ll_BallTree(tree = coast_cell_tree, cells = nurdle_query)

# Assign Region Labels
regions = gom_masks.group_source_region_map.values[idx_coast_mask]
nurdle_regions = regions[cell_beaching_inds].astype(int)
df_all_nurdles['region'] = group_source_region_names[(nurdle_regions - 1)]

# Build Nurdle Stats Table
nurdle_stats = df_all_nurdles.groupby('region').nurdle_count.agg(['count', 'sum', 'mean'])
nurdle_stats['count_perc'] = 100 * nurdle_stats['count'] / nurdle_stats['count'].sum()
nurdle_stats['sum_perc'] = 100 *nurdle_stats['sum'] / nurdle_stats['sum'].sum()
nurdle_stats.loc['Total']= nurdle_stats.sum()

nurdle_stats = nurdle_stats[['count', 'count_perc', 'sum', 'sum_perc', 'mean']]
nurdle_stats = nurdle_stats.rename(columns={"count": "survey_count", 'count_perc':'survey_perc', 
                             'sum':'total_nurdle_count', 'sum_perc': 'total_nurdle_perc', 
                             'mean' : 'mean_nurdle_count'})
nurdle_stats = nurdle_stats.round(acc-1)

nurdle_stats.to_csv(outputgraphsDir + 'nurdle_stats.csv')

Nurdle Beaching Month

# Get Nurdle Release Region by Month 
release_region_month = dict(df_all_nurdles.groupby([df_all_nurdles.patrol_date.dt.month, df_all_nurdles.region]).nurdle_count.sum())

# Prepare Data for Plotting
nurdle_release_region_names = ['Texas', 'Louisiana', 'Mississippi', 'Alabama', 'Florida']
nurdle_release_region_names_idx = {'Texas': 0, 'Louisiana': 1, 'Mississippi': 2, 'Alabama': 3, 'Florida': 4}


nurdle_months = {(i+1): np.zeros(len(nurdle_release_region_names_idx)) for i in range(np.shape(Fig13_vals)[1])}
for month,region in release_region_month.keys():
    val = release_region_month[(month, region)]
    nurdle_months[month][nurdle_release_region_names_idx[region]] += val
    
# Normalize
nurdle_months = {i: nurdle_months[i] / np.sum(nurdle_months[i]) for i in nurdle_months.keys()}

category_names_fig2 = nurdle_release_region_names
# since vals are 0.999, I have just adjusted them to exactly one, and the plot looks better this way 
results_fig2 = nurdle_months

nurdle_colors = []
for name in nurdle_release_region_names:
    nurdle_colors.append(region_colors[name])

def survey(results, category_names):
    """
    Parameters
    ----------
    results : dict
        A mapping from question labels to a list of answers per category.
        It is assumed all lists contain the same number of entries and that
        it matches the length of *category_names*.
    category_names : list of str
        The category labels.
    """
    labels = list(results.keys())
    data = np.array(list(results.values()))
    data_cum = data.cumsum(axis=1)
    category_colors = nurdle_colors

    fig, ax = plt.subplots(figsize=(9.2, 5))
    ax.invert_yaxis()
    ax.yaxis.set_visible(True)
    ax.set_ylim(0, np.sum(data, axis=1).max())
    ax.set_ylabel("Beaching Contribution %")
    ax.set_xlabel('Beaching Month')


    for i, (colname, color) in enumerate(zip(category_names, category_colors)):
        widths = data[:, i]
        starts = data_cum[:, i] - widths
        rects = ax.bar(labels, widths, bottom=starts,width = 0.9, label=colname, color=color, edgecolor='white',)

    legend = ax.legend(ncols=1, bbox_to_anchor=(1.25, 0.5), loc='right', fontsize='medium')

    return fig, ax, legend


fig, ax, legend = survey(results_fig2, category_names_fig2)

labels = ['Jan','Feb','Mar','Apr','May','Jun','Jul','Aug','Sept','Oct', 'Nov', 'Dec']
x = [i for i in range(1, len(labels) + 1)]
plt.xticks(x, labels, rotation=45)
plt.savefig(outputgraphsDir + 'nurdle_release_month.png')
plt.close()
